SpaceStore: 将金币存到打印舱

+++ SpaceStore/SellButtons/ElementSellButton.cs
@@ -12,9 +12,11 @@ namespace SpaceStore.SellButtons {
         }
 
         public override void Sell() {
+            if (StaticVars.coinSaver == null) { return; }
             if (coin == 0) { CountPrice(); }
             if (coin < 0) { return; }
-            StaticVars.AddCoin(coin);
+            //StaticVars.AddCoin(coin);
+            StaticVars.coinSaver.AddCoin(coin);
             primaryElement.gameObject.DeleteObject();
             base.Sell();
         }
